Introduce installation from GitHub releases
As of version 4.x, the ccache GitHub releases include
statically-compiled binaries for various platforms. This installation
method is much quicker and simpler than building by source or installing
from the system's package manager, but may not be appropriate in all
scenarios. For example, no other resources (such as the documentation)
are installed.